requirements.txt
pikepdf>=8.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0